fix bug in opacity colour map
git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/GraphViz/Trunk@41726 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/plugins/graphviz/colourMaps/ColourMapOpacity.java
@@ -12,12 +12,12 @@ public class ColourMapOpacity extends ColourMap {
 	public Color colour(long weight, long maxWeight) {
 		double opacity = weight / (double) maxWeight;
 		Color colour = base.colour(weight, maxWeight);
-		return new Color(colour.getRed(), colour.getGreen(), colour.getBlue(), Math.round(opacity * 255));
+		return new Color(colour.getRed(), colour.getGreen(), colour.getBlue(), (int) Math.round(opacity * 255));
 	}
 
 	public Color colour(double opacity) {
 		Color colour = base.colour(opacity);
-		return new Color(colour.getRed(), colour.getGreen(), colour.getBlue(), Math.round(opacity * 255));
+		return new Color(colour.getRed(), colour.getGreen(), colour.getBlue(), (int) Math.round(opacity * 255));
 	}
 
 }
